Melhorando estrutura
