+++ package.json
@@ -23,6 +23,7 @@
     "dependencies": {
         "async-redis": "^1.1.7",
         "axios": "^0.18.0",
+        "body-parser": "^1.18.3",
         "dotenv": "^7.0.0",
         "express": "^4.16.4"
     }
